+++ src/main/resources/templates/login_form.html
@@ -21,7 +21,7 @@
   </div>
 </form>
 <div>
-  <a th:href="@{/users/kakao/login}">카카오 로그인</a>
+  <a th:href="@{/oauth/kakao/login}">카카오 로그인</a>
 </div>
 <script th:inline="javascript">
   document.getElementById('loginForm').addEventListener("submit", async function (event) {
